requirements.txt
openai==1.75.0
pydantic_settings==2.9.1
python-dotenv==1.1.0
pydantic==2.6.1
typing-extensions>=4.8.0
